Update dotnet-install scripts to support ubuntu 18 and alpine

+++ files/KoreBuild/scripts/dotnet-install.ps1
@@ -73,6 +73,8 @@
 .PARAMETER SkipNonVersionedFiles
     Default: false
     Skips installing non-versioned files if they already exist, such as dotnet.exe.
+.PARAMETER NoCdn
+    Disable downloading from the Azure CDN, and use the uncached feed directly.
 #>
 [cmdletbinding()]
 param(
@@ -91,12 +93,17 @@ param(
    [string]$FeedCredential,
    [string]$ProxyAddress,
    [switch]$ProxyUseDefaultCredentials,
-   [switch]$SkipNonVersionedFiles
+   [switch]$SkipNonVersionedFiles,
+   [switch]$NoCdn
 )
 
 Set-StrictMode -Version Latest
 $ErrorActionPreference="Stop"
 $ProgressPreference="SilentlyContinue"
+
+if ($NoCdn) {
+    $AzureFeed = $UncachedFeed
+}
 
 $BinFolderRelativePath=""
 
@@ -461,7 +468,7 @@ function Extract-Dotnet-Package([string]$ZipPath, [string]$OutPath) {
 }
 
 function DownloadFile([Uri]$Uri, [string]$OutPath) {
-    if ($Uri -like "file://*") {
+    if ($Uri -notlike "http*") {
         Say-Verbose "Copying file from $Uri to $OutPath"
         Copy-Item $Uri.AbsolutePath $OutPath
         return
@@ -528,6 +535,7 @@ else {
     throw "Invalid value for `$Runtime"
 }
 
+#  Check if the SDK version is already installed.
 $isAssetInstalled = Is-Dotnet-Package-Installed -InstallRoot $InstallRoot -RelativePathToPackage $dotnetPackageRelativePath -SpecificVersion $SpecificVersion
 if ($isAssetInstalled) {
     Say "$assetName version $SpecificVersion is already installed."
@@ -566,6 +574,12 @@ catch {
 
 Say "Extracting zip from $DownloadLink"
 Extract-Dotnet-Package -ZipPath $ZipPath -OutPath $InstallRoot
+
+#  Check if the SDK version is now installed; if not, fail the installation.
+$isAssetInstalled = Is-Dotnet-Package-Installed -InstallRoot $InstallRoot -RelativePathToPackage $dotnetPackageRelativePath -SpecificVersion $SpecificVersion
+if (!$isAssetInstalled) {
+    throw "$assetName version $SpecificVersion failed to install with an unknown error."
+}
 
 Remove-Item $ZipPath
